Added Geometry Interfaces.
These are part of a new spec which will give more power to graphics programmers. I believe there is intent to create the ability to pass DOMMatrix instances to Elements for increased performance. Currently, the only way is to set a CSS string in the `style` attribute of an Element, which incurs performance loss due to string manipulation (compared to the numerical approach with DOMMatrix).