Develop Custom Elements
Pleeeease, this is an awesome change for the web. http://aframe.io is the current reference-example of the power that this API brings to web developers, allowing them to create HTML APIs for other devs to use. I'm also working on an HTML API using `document.registerElement` for the engine that we're building at http://infamous.io (name subject to change).