Change status of screencapture to 'In Development'
It has recently been confirmed that an underlying dependency
to getDisplayMedia (namely the CapturePicker) is not present
on Windows Home installations, but instead only on Windows
Pro and Enterprise.

Since there is no reliable way of detecting if a client is
differentiating between Windows Home and Pro/Enterprise users
without actually triggering screenshare and catching the
error, if any is thrown, it can be considered unsupported.